[#3] split files by functionality
